refactor(msteams): move configuration types into msteams package
